aanpassing van frontend en backend
